Refactor the query parsing logic out of the main extension
This commit refactors the query parsing logic, into a dedicated file
that can be unit tested directly using node. The main extension file is
now responsible of getting the data from the DOM and passing a set of
valid filters down to the `getQuery` function.

Basic tests were added to the query translator and can be executed via:
```
npm test
```

Changes to the manifest are needed to first inject the file that
contains the parsing logic before `inject.js`.